requirements.txt
streamlit>=1.34.0
google-generativeai>=0.5.0
gTTS
python-dotenv
